fix: delete message after get photos

+++ src/bot/conversations/create-ticket/create-ticket.ts
@@ -52,7 +52,9 @@ export const createTicketConversation = (container: Container) =>
       ? getPriority({ ...conversationProperties, ctx: categoryCtx })
       : [undefined, ctx]);
 
-    await priorityCtx.deleteMessage();
+    if (isManagerUser) {
+      await priorityCtx.deleteMessage();
+    }
 
     const [photoUrs, photosCtx] = await getPhotos({
       ...conversationProperties,
